Fix issue with basic_consume paremter order between v0.13.1 and v1.0.0+
In version 0.13.1 the adapters/blocking_connection.py defined basic_consume with parameters consumer_callback and queue in that order. In version 1.0.0 they changed the parameters to queue and on_message_callback with this new order.

This fix uses named parameters to be compatible with version 1.0.0+  but breaks compatibility with version 0.13.1 or earlier